Fixed some compile errors with gcc 4.1.2
git-svn-id: https://svn.svp-home.org/MGSim/trunk@509 e97e5017-a994-416e-809d-76780e9f78db

+++ functions.h
@@ -22,7 +22,7 @@ public:
     }
 
     ArbitratedFunction(Kernel& kernel)
-        : m_kernel(kernel), m_priority(SIZE_MAX), m_component(NULL)
+        : m_kernel(kernel), m_priority(INT_MAX), m_component(NULL)
     {
         m_kernel.registerFunction(*this);
     }
